- Move Lattica's submission to submission_remote subdir - Add upload-file steps to harness - Change harness to directly invoke py files
